# PERF: Use Categorical.equals in CategoricalIndex.equals