@pep8speaks 2018-11-30
Hello @TomAugspurger! Thanks for submitting the PR.

 - There are no PEP8 issues in the file [`pandas/core/indexes/category.py`](https://github.com/TomAugspurger/pandas/blob/09e07d2ae26008d203bc84e4721fa83ff484e662/pandas/core/indexes/category.py) !